Making the MongoDB sink strict on target type conversion for properties and keys
